Create project template with basic Makefile
From the outset I'm guessing that most of these projects are going to be
similar in structure. This isn't a super-complicated "let's-make-an-OS"
class, this is a data structures class. That said, there'll be a src/
dir, a include/ dir, and maybe a README. So I made a template that
contains all of those things, including a basic Makefile.

Speaking of...it's VERY basic. The coolest thing it can do is deploy
files to GL, and even that's not for sure gonna work once it hits
production. We'll see. Anywhere, since the "compilation command" is
supposed to change per-project, I figured I'd leave every target but
`deploy` alone and allow the user (me) to implement it later.