Add ignore rule for script dump files
We have to do this annoying thing where we "record" our compilation and
debugging procedure using the `script` command. Wonderful security,
that. Anyway, I don't want the files that `script` generates committed
to the repo, because then EVERYONE could see just how shitty of a
developer I am sometimes!

There will be more lines to come in the .gitignores. I'm trying to keep
them as segmented as possible, e.g. files to be ignored in proj1 ONLY
are configured to be ignored in a separate .gitignore inside proj1.