Add run command macro and remove echos
I realized that some projects are probably going to have command-line
parameters, so it would make sense to allow for a project-specific run
command just as I allowed for a project-specific build command. Maybe
I'll eventually implement a project-specific clean command.

I also removed the echo statements because really they're unnecessary (I
mean, I know which target I just ran) and they often get lost in the sea
of text that is printed as compilations and executions occur.

(plus they were showing up in the typescript which might embarass me in
front of the TA's!)