Got basic inner circular buffer working
Diagnosed a segfault being caused by not properly initializing the
data buffer member, but realized they wanted me to use the `new` keyword
instead of manually allocating it. So all's well on that front.

Also wrote a short test class (that really only applies to this commit,
as from here on out I'll be trying to move to the provided test classes)
that test the enqueue/dequeue functionality of the ICB.