Change project subsystem type to console
This change affects the project's behavior in Visual Studio only. The
project was created in Visual Studio as an Empty Project (to avoid
needing to include stdafx.h, et. al). When this is done, Visual Studio
leaves the Subsystem configuration unset, causing the debugger window to
close immediately after the program returns, rather than staying open to
allow the developer to examine its output.

By changing the subsystem type to Console, the debugger window will stay
open after it is launched and the program returns.