Create Makefile with various build tasks (broken)
I don't know why, but the 'build' target isn't working correctly yet. I
think it may have something to do with the fact that I created the
project in Visual Studio first and *then* added a Makefile. VS may be
setting some weird setting or doing something funky with pathing that's
causing the Makefile to get confused and not build. Basically, the
Makefile can't find a main() even though it's defined in driver.cpp.

I spent an hour or two debugging it, but it was taking too much time so
I decided to cut losses and keep using Visual Studio for coding and
building until I get a little further down the road. If I make a lot of
progress I'll revisit the Makefile project.